agents/chat-analysis-agent/pyproject.toml
[project]
name = "chat-analysis-agent"
version = "0.1.0"
description = "Agent for chat analysis system"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "google-adk[a2a]>=1.16.0",
    "litellm>=1.78.0",
    "loguru>=0.7.3",
    "fastmcp>=0.1.0",
]
